GoogleRecaptcha/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-recaptcha==3.0.0
Pillow==9.4.0
sqlparse==0.4.3
